- KOMI implemented http://senseis.xmp.net/?Komi - and fix the problem when android back button pressed

+++ app/src/main/java/beothorn/github/com/toroidalgo/ScoreListener.java
@@ -1,5 +1,5 @@
 package beothorn.github.com.toroidalgo;
 
 public interface ScoreListener {
-    void setScore(int blackScore, int whiteScore);
+    void setScore(int blackScore, float whiteScore);
 }
