+++ app/src/main/java/beothorn/github/com/toroidalgo/go/impl/logic/BoardListener.java
@@ -2,7 +2,7 @@ package beothorn.github.com.toroidalgo.go.impl.logic;
 
 public interface BoardListener {
 
-	void updateScore(int _blackScore, int _whiteScore);
+	void updateScore(int _blackScore, float _whiteScore);
 	void nextToPlay(GoBoard.StoneColor _nextToPlay);
 	void nextToPlayOnPass(GoBoard.StoneColor _nextToPlay);
